requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml
langchain-unstructured[local]
unstructured-client 
unstructured
libmagic
arxiv 
pymupdf
wikipedia
youtube_search
langchain-tavily
langgraph
mlflow
langchain_google_genai
langchain_huggingface
chromadb
langchain-chroma>=0.1.2
protobuf==3.20.*
duckduckgo-search
